PI-4011-signed removed CaseService to fix duplication
